g_packages/official_py_docker/docker/jammit/requirements.txt
numpy
pandas
matplotlib
scipy
ansicolors